supprimer un patient est ajoutee

+++ src/main/java/ma/enset/hospital_system/repository/PatientRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientRepository  extends JpaRepository<Patient, Integer> {
+public interface PatientRepository  extends JpaRepository<Patient, Long> {
 
     Page<Patient> findByNomContains(String keyword , Pageable pageable);
 }
